54_transferlearning_freezing/Akhund-Human/requirements.txt
tensorflow
matplotlib.pyplot
sklearn
numpy
wandb
telebot
scipy
Pillow
opencv-python
dlib